moved integration test to integration test module
